Move around fetchers code.
